moving Static one level up, import, style.js changes
