+++ src/components/Static/index.js
@@ -1,0 +1,4 @@
+export { default as AppDescription } from './AppDescription';
+export { default as LocalConversionDescription } from './LocalConversionDescription';
+export { default as EmptyConversionAttempt } from './EmptyConversionAttempt';
+export { default as LoadingAnimation } from './LoadingAnimation';
