+++ src/components/ZipMarkdownToHtml/style.js
@@ -13,7 +13,7 @@ export const useStyles = makeStyles((theme) => ({
   },
 
   backdrop: {
-    zIndex: theme.zIndex.drawer + 1,
+    zIndex: theme.zIndex.drawer,
     color: theme.palette.grey[100],
   },
   paper: {
